goal4/freshrank/requirements.txt
fastapi
uvicorn
pydantic
pyyaml
loguru
pandas
numpy
regex
dateparser
pdfplumber
pymupdf
camelot-py[cv]
python-dateutil
scikit-learn
pytest
httpx
